% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context_example.R
\name{example_crossline}
\alias{example_crossline}
\title{Example function calling abline}
\usage{
example_crossline(...)
}
\arguments{
\item{...}{arguments passed to \code{abline}}
}
\description{
Example function calling abline
}
