% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context_example.R
\name{example_one_plot}
\alias{example_one_plot}
\title{Example function calling example_crossline twice}
\usage{
example_one_plot(..., h = 5, v = 5)
}
\arguments{
\item{...}{arguments passed to \code{plot}}

\item{h}{position of horizontal line}

\item{v}{position of vertical line}
}
\description{
Example function calling example_crossline twice
}
