% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context_example.R
\name{example_three_plots}
\alias{example_three_plots}
\title{Example function calling example_one_plot three times}
\usage{
example_three_plots()
}
\description{
Example function calling example_one_plot three times
}
