% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context_main.R
\name{getContext}
\alias{getContext}
\title{Get the context tree (currently from a global variable "CONTEXT\code{n}")}
\usage{
getContext(number = getDefault("getContext", "number"))
}
\arguments{
\item{number}{Context number (see \code{\link{initContext}})}
}
\value{
a list of function calls
}
\description{
Get the context tree (currently from a global variable "CONTEXT\code{n}")
}
