% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context_log.R
\name{getFunctionCall}
\alias{getFunctionCall}
\title{Get a function call from the context tree}
\usage{
getFunctionCall(i, number = getDefault("getFunctionCall", "number"))
}
\arguments{
\item{i}{Index of the function call as it is stored in the context}

\item{number}{Number of the context (see \code{\link{initContext}})}
}
\description{
Get a function call from the context tree
}
