% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context_main.R
\name{initContext}
\alias{initContext}
\title{Initialise the variable storing the full function call context}
\usage{
initContext(number = getDefault("initContext", "number"))
}
\arguments{
\item{number}{Number of the context. You may use different contexts that are
numbered. The default context number is 1.}
}
\description{
Initialise the variable storing the full function call context
}
