% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context_main.R
\name{printContext}
\alias{printContext}
\title{Print the context of a function call}
\usage{
printContext(number = getDefault("printContext", "number"), indent = "  ")
}
\arguments{
\item{number}{Number of the context to print (see \code{\link{initContext}})}

\item{indent}{string used to indent nested function calls in the output}
}
\description{
Print the context of a function call
}
